+++ web/src/pages/assessment-page/assessment-summary.tsx
@@ -129,7 +129,7 @@ const AssessmentSummary = () => {
                             size='large'
                             onClick={() => quitApp()}
                         >
-                            Quit
+                            Quit Visreg
                         </Button>
                     </Grid>
                 </Grid>